add version

+++ geodk.py
@@ -1,13 +1,14 @@
 import pickle
 import requests
-with open('geo_lookups.pkl','wb') as f:
-    f.write(requests.get('https://github.com/snorre87/dk_geo/raw/main/geo_lookups.pkl').content)
-    f.close()
-
+import os
+if not os.path.isfile('geo_lookups.pkl'):
+    with open('geo_lookups.pkl','wb') as f:
+        f.write(requests.get('https://github.com/snorre87/dk_geo/raw/main/geo_lookups.pkl').content)
+        f.close()
 pn2pnum,pnr2kom,p2kom,kom2reg,kom2reg,reg2reg,sogn2zip,z2geo,sogn2geo = pickle.load(open('geo_lookups.pkl','rb'))
 pnum2pn = {j:i for i,j in pn2pnum.items()}
 final_regs = set(reg2reg.values())
-
+version = 0.1
 def get_geo_info(geoname):
     info = {}
     assert type(geoname)==str, 'Input has to be string'
